Convert to the PyQt5 Framework and Remove Extra Files
Several files that should be the output of .deb generation
were included in the repository.  Removing those files for
now because they just make extra clutter.  Will work on
a script to generate those files to build .deb package
and maybe .rpm later.

The WX framework is old, and it seems to make more sense to
use QT to create the GUI for this program instead.  Changed
code that generates GUI components to use PyQt5 instead of
WX.