Remove Excess Wordiness and Some Code Cleanup
On the first pass converting this program to use PyQt, I imported
the QtWidgets, QtGui, and QtCore modules from PyQt.  While this
worked fine it required explicit references to those modules in
order to access any PyQt classes.  This made for excessively
wordy code with a lot of repetition.  Changed the import call
to import all classes from PyQt.QtWidgets, etc., and removed
explicit references to those modules in the code.

Changed Help/About Dialog to display FSF address in a more
readable way.

Removed code that resized objects in window when the window is
resized.  This is not necessary because the QT layout managers
handle that just fine.

Created a MyPanel subclass for MyApp, which creates a widget with
a layout to contain another widget.  This makes it easier to add
a nice looking panel layout to add to a tab or a frame in a way
that it looks like it belongs inside the tab or frame.