Fix Up the Indentation in the Python Source
The indentation is inconsistent in the source files, which makes
updating the Python code frustrating as the spaces have to be
entered manually for each new line.  This change updates the
indentation so each level of indentation is exactly 4 spaces, and
vim settings are added to the top of each file to set the editor's
tab stops to 4 spaces to make editing the file easier.